requirements.txt
fastapi
uvicorn
sqlmode
sqlalchemy
pytest
httpx
requests
